+++ CHANGELOG.md
@@ -9,9 +9,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - convenient IAM Role for creating Flow Logs
 - support Amazon-assigned IPv6 (disabled by default)
 - demonstrate cloud-init in example-service
+- vpn-connection now supports Transit Gateway (as well as VPN Gateway)
 
 ### Changed
 - update example-service to Amazon Linux 2, t3.nano
+- vpn-connection CloudWatch Alarm uses native metrics instead of old custom "VPNStatus" metrics
+- SNS topics for VPN monitoring alerts are now per-region
+
+### Removed
+- old custom "VPNStatus" metrics solution (Lambda deployed via CloudFormation)